refactor: exception과 config 패키지 정리
