+++ src/main/java/wooteco/subway/WebConfig.java
@@ -1,4 +1,4 @@
-package wooteco.subway.config;
+package wooteco.subway;
 
 import com.google.common.net.HttpHeaders;
 import org.springframework.context.annotation.Configuration;
