+++ src/main/java/wooteco/subway/exception/AuthorizationException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.auth.application;
+package wooteco.subway.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
